+++ src/features/wallet/screens/Wallet/components/AccountValue/constants.tsx
@@ -7,31 +7,31 @@ import { WalletRoute } from '@app/features/wallet/navigation/constants';
 
 export const accountValueButtonsList: AccountValueButtonsList[] = [
   {
-    button: <AppIcon name="ArrowDown" color="white" />,
+    Button: ({ color }) => <AppIcon name="ArrowDown" color={color} />,
     title: 'Receive',
     route: WalletRoute.Deposit,
     id: AccountValueButtonsId.Deposit,
   },
   {
-    button: <AppIcon name="ExternalLink" color="white" />,
+    Button: ({ color }) => <AppIcon name="ExternalLink" color={color} />,
     title: 'Send',
     route: WalletRoute.Withdraw,
     id: AccountValueButtonsId.Withdraw,
   },
   {
-    button: <AppIcon name="Money" color="white" />,
+    Button: ({ color }) => <AppIcon name="Money" color={color} />,
     title: 'Buy',
     route: WalletRoute.WithdrawDetails,
     id: AccountValueButtonsId.Buy,
   },
   {
-    button: <AppIcon name="Switch" color="white" />,
+    Button: ({ color }) => <AppIcon name="Switch" color={color} />,
     title: 'Exchange',
     route: WalletRoute.DepositDetails,
     id: AccountValueButtonsId.Swap,
   },
   {
-    button: <AppIcon name="List" color="white" />,
+    Button: ({ color }) => <AppIcon name="List" color={color} />,
     title: 'History',
     route: WalletRoute.History,
     id: AccountValueButtonsId.History,